rename env_config.py to utils.py and change imports
